Rewrote open_repos to use conditional gets.
ETag and Last-Modified HTTP headers are now used to see if a repo has changed
since the last time the client read it.  This cuts down on bandwidth and
processing time.

To do this, an extra database table is created, acting as an index of all
remote repositories.  It assumes that the repo's URL is its unique identifier.
update_remote needs rewriting to use the URL as the table name for each repo.
Then, database_query can use this index table to know the table names for the
remote repositories.