Added support for repo spec version 1.1.
Version 1.1 of the spec adds a requirement for each PND's MD5 hash to be
included.  This value will now be included in all the remote repository tables.
However, I decided to maintain backwards compatibility with version 1.0, as
well as fault tolerance; therefore, update_remote will not complain if a v1.1
repo misses the hash, nor if v1.0 includes it.