Table names for a remote repo is now the repo's URL.
This guarantees uniqueness and allows database queries to check the repo index
for the table name.  In the process, I made update_remote add the repo's name
to the index, and moved the ETag/Last-Modified update to the same place.  This
guarantees that the repo's table is completely updated before storing ETag and
Last-Modified.