Added PND searchpath definition to options.
Function options.get_searchpath returns the list of paths in which to search
for PND files.  If it encounters "default", it interprets that as the
searchpath defined in libpnd's config files, and gets it accordingly.

Also added corresponding tests.  In order to make the tests work without having
to make system-level changes, testfiles folder was renamed to testdata, and
conf files were moved into a subfolder called conf.  This is because libpnd
automatically searches for conf files in "./testdata/conf".  However, this
means tests must always be run from folder test, unless libpnd is installed
system-wide (as it would be on a Pandora).