Modified options.get_locale to use "default" key.
The system's default locale is now accessible in the config file by specifying
the string "default".  This is the same behaviour as is in get_searchpath.
Correspondingly, a "locale" entry is now required in the config; it will fail
rather than fall back on the system default if the entry is not found.  Tests
updated accordingly.