Wrote update_local.  With that, database_update is about done.
To achieve this, update_local_file was also created.  It currently only handles
the first app listed in a PND (since that's what milkshake's repo does for
now), but it should be easily extended to accommodate multiple apps.

Since these functions rely heavily on libpnd, they're susceptible to segfaults.
Some checks are done to prevent them, but may need more.

A proper test for update_local was created.  Perhaps a test for
update_local_file is also in order, though it is covered by update_local.
Tests for bad conditions (that could cause segfaults) are also in order.