Most work needed to find updates is done.
Functions in database_query can fetch info on all available versions of a
package and sort it from newest to oldest, so the rest is just bookkeeping.
But I'm thinking of using an OO interface for database_query instead; it could
allow for better querying without having to be familiar with the returned data
structures.  But I want to make sure code so far doesn't get lost.  See future
commits to find out what I choose to do.