Updated to repo spec 2.0, dropping older versions.
Changed the database table layout to match.  Rewrote some of the update_local
code to take advantage of this, but libpnd can't yet read all the fancy things
in a PXML.  In particular, it can't read the package element, which means it
might not be able to make proper comparisons with data in the repo.  Worked
around by using the first application in lieu of package data, which is
probably fine for many PNDs.

Will want to have a second table for every repository (including local),
listing all the applications found in PNDs, but that is for the future.