Reworked database_query, renamed it to packages as a result.
Decided on a much nicer object-oriented approach to querying and handling
packages.  Implemented some of the important query-related functions, but it
will eventually handle installation and removal as well, so the module was
renamed to reflect its role.  Updated tests, of course, which included updating
the example repo file to a 2.0 version (thanks to milkshake).