Replaced supertest.pnd with fulltest.pnd, added more tests.
fulltest.pnd is a slightly-modified version of full-sample_PXML.xml from the
libpnd docs.  I named it .pnd so it would be read as a PND file for parsing.
It uses every feature available to a PXML, so is a good test; I correspondingly
updated testUpdateLocal, but I obviously choke on it since I don't yet parse
package elements.  Once I add that in, I'll try to apply tests on fulltest.pnd
more widely.