Now read package element by extracting PXML.
Who knows when libpnd will get full support for the latest PXML spec, so I
worked around it.  I use libpnd to extract the PXML, then parse the bits that
it can't.  I continue to use libpnd to parse the bits that it can, though maybe
I should drop that and do solely XML parsing.

Note that license elements aren't yet parsed.  I'll perhaps have to add that in
later.