Fixed segfault on systems without a locale specified.
The Plugapps chroot on my Touch Book had no locale, so the default locale was
appearing as "None".  libpnd doesn't like that.  So options.get_locale() now
watches for that and removes it if it appears.  The corresponding test was
updated; I can't replicate the condition on a machine that does have a locale,
but the test now works even on machines that don't have a locale.