Created wrapper to preload libstdc++.so.
I learned before that libpnd needed libstdc++ linked into it in order to start,
but I assumed the same problem did not exist on Pandoras.  Well, it does!  I
made a wrapper script around the PNDstore GUI executable to preload libstdc++,
but I'll have to submit a bug report on it so it can be fixed by next release.