Removed "Not Null" requirement for title and description.
IsSuE reports an error that seems to be coming from one of his PNDs not having
a description.  Descriptions and titles are mandatory in a PXML, but that
doesn't mean they're always there.  So I made the database more permissive to
allow such PNDs.  Even though they're poorly-formed, I can still handle a
missing description.