If a repo can't be reached, gives a warning and keeps going.
The fix for last commit's test update (which itself needed an update for a
little mixup, but it's good, really).  With that, all the important errors that
an end user might see are turned into warnings, letting the code chug along as
best it can.  Doubtless there will be more in the future, though, especially if
multiple repos come along.