Handle hopefully all issues related to table existence.
The get_all and get_all_local convenience functions in packages would still
error if the local and index tables weren't in place.  Then I found another
issue with tables being listed in the index, but not being created if there was
an HTTP error.  These now all have tests and fixes in place.