Removed MD5 calculation on already installed packages.
I want update_local to be as fast as possible, and calculating MD5 requires
reading the whole file.  MD5 of local files isn't used; I thought I could use
it to determine if local and remote versions are identical, but there's no need
for that yet.  So, save time by commenting it out (also associated tests).