Basic threading of important operations in GUI.
Update, install, and upgrade now all run in the background.  A statusbar was
added to the window to notify on progress and warnings.  But only one such
operations thread is allowed to run at a time, since modifications to the
database can't be made concurrently; an error appears if you try (but the db
can't be corrupted in the process, thankfully).