Failed upgrades should not remove a PND without replacing it.
Though it was never reported, the package upgrade method could have easily
removed a PND without replacing it if something in the upgrade process failed
(like the download).  Now, the old version is instead stored temporarily until
the new version is completely in place; if failure occurs at any point, the old
version is moved back to where it was.