Created function to search for installed packages.
Finds all packages with the specified value contained in the specified column.
Is particularly designed for searching within the applications and categories
columns, which contain semicolon-delimited lists.  The function will only match
those packages that contain the value as a full element of the list, not just a
substring.  But substring searching can still be performed by using SQL
wildcards (notable %) in the value to be found.