Only one Package object can exist for each given ID.
This ensures that multiple requests for the same package will always return the
same object.  This is beneficial for package searching, as it allows duplicates
to be removed with a simple call to "set".  It also means that installing or
upgrading on one instance will not cause other instances to become out-of-sync
with the database (as there are no other instances!).  Is this a variant of the
Singleton pattern?