Implemented basic features of repo v3.
Table columns were rearranged to match the order they appear in the spec (just
for readability).  Extra columns were added to hold the extra data that's
available.  And icon_cache was removed since I don't use it yet and probably
won't even once icon downloading is implemented.

Also, the JSON package parsing was moved out into a separate function in
anticipation of the updates-only URI.  In the process, more robust parsing was
implemented; a single failed package will not stop the rest of the repo from
being parsed, and only a missing ID or URI should be able to take it down.
Also now append version type to the end of the version string if it's not
"release", but I may rework that just so versions look nicer when displayed.