Started on repo v3 improved updating.
Implemented the first half of the update procedure: full and conditional
updates.  Right now, it effectively works the same as it did before, but is now
in a place to implement the updates-only URL.

Note that, in database_update, open_repos has been removed entirely, and
update_remote_url has basically been rewritten.  It now takes a string URL
rather than an open file object, and can be forced to perform a full update
without conditional gets.  But it's not done yet!