Fixed bug causing remote update failure if first remote update fails.
If the first remote update for a repo fails, then the repo's last_full_update
would be set to None, causing subsequent attempts to update to fail with a
TypeError as it tries to convert None to an int.  This fix ensures that
last_full_update is set to 0 on a repo's first update.  Also, a corresponding
test was added.