Moved error printing in GUI to a subfunction.
There were four places that would show errors on the statusbar, mostly with the
same logic.  This was moved to a method of the PNDstore class.  This method was
also changed from the original code in that it also prints the errors to
stdout, thereby ensuring that they'll end up in the .out file when run as PND.